% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadButtonApp.R
\name{downloadButtonApp}
\alias{downloadButtonApp}
\alias{downloadButtonServer}
\alias{downloadButtonInput}
\alias{downloadButtonUI}
\alias{downloadButtonOutput}
\title{Download App with Buttons}
\usage{
downloadButtonApp(id)

downloadButtonServer(id, download_list)

downloadButtonInput(id)

downloadButtonUI(id)

downloadButtonOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{download_list}{reactiveValues object}
}
\description{
Download App with Buttons
}
