% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeApp.R
\name{mergeApp}
\alias{mergeApp}
\alias{mergeServer}
\alias{mergeUI}
\title{Merge App}
\usage{
mergeApp(id)

mergeServer(id, primary_list, secondary_list)

mergeUI(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{primary_list, secondary_list}{reactiveValues objects}
}
\description{
Merge App
}
