% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadApp.R
\name{downloadApp}
\alias{downloadApp}
\alias{downloadServer}
\alias{downloadInput}
\alias{downloadUI}
\alias{downloadOutput}
\alias{downloadShow}
\title{Download App}
\usage{
downloadApp(id)

downloadServer(id, download_list)

downloadInput(id)

downloadUI(id)

downloadOutput(id)

downloadShow(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{download_list}{reactiveValues object}
}
\description{
Download App
}
